+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#18](https://github.com/LaxarJS/ax-input-control/issues/18): fixed updates to tooltip message
+
 
 ## v2.0.0-alpha.6